fix some code review problems
